- Separate AlertState & ActionSheetState from the SwiftUI extension - Create UIKit extensions for Alerts
