+++ Sources/ComposableArchitecture/UI/AlertState.swift
@@ -1,3 +1,13 @@
+//
+//  AlertState.swift
+//  ComposableArchitecture
+//
+//  Created by Eric Blachère on 11/08/2020.
+//  Copyright © 2020 Bioche. All rights reserved.
+//
+
+import Foundation
+
 /// A data type that describes the state of an alert that can be shown to the user. The `Action`
 /// generic is the type of actions that can be sent from tapping on a button in the alert.
 ///
@@ -159,81 +169,3 @@ extension AlertState: Equatable where Action: Equatable {}
 extension AlertState: Hashable where Action: Hashable {}
 extension AlertState.Button: Equatable where Action: Equatable {}
 extension AlertState.Button: Hashable where Action: Hashable {}
-
-extension AlertState: TCAIdentifiable where Action: Hashable {
-  public var id: Self { self }
-}
-
-#if canImport(Combine)
-import SwiftUI
-
-@available(iOS 13, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
-extension View {
-  /// Displays an alert when then store's state becomes non-`nil`, and dismisses it when it becomes
-  /// `nil`.
-  ///
-  /// - Parameters:
-  ///   - store: A store that describes if the alert is shown or dismissed.
-  ///   - dismissal: An action to send when the alert is dismissed through non-user actions, such
-  ///     as when an alert is automatically dismissed by the system.
-  public func alert<Action>(
-    _ store: Store<AlertState<Action>?, Action>,
-    dismiss: Action
-  ) -> some View {
-
-    let viewStore = ViewStore(store, removeDuplicates: { ($0 == nil) != ($1 == nil) })
-    return self.alert(
-      isPresented: Binding(
-        get: { viewStore.state != nil },
-        set: {
-          guard !$0 else { return }
-          viewStore.send(dismiss)
-        }),
-      content: { viewStore.state?.toSwiftUI(send: viewStore.send) ?? Alert(title: Text("")) }
-    )
-  }
-}
-
-extension AlertState: Identifiable where Action: Hashable { }
-
-
-@available(iOS 13, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
-extension AlertState.Button {
-  func toSwiftUI(send: @escaping (Action) -> Void) -> SwiftUI.Alert.Button {
-    let action = { if let action = self.action { send(action) } }
-    switch self.type {
-    case let .cancel(.some(label)):
-      return .cancel(Text(label), action: action)
-    case .cancel(.none):
-      return .cancel(action)
-    case let .default(label):
-      return .default(Text(label), action: action)
-    case let .destructive(label):
-      return .destructive(Text(label), action: action)
-    }
-  }
-}
-
-@available(iOS 13, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
-extension AlertState {
-  fileprivate func toSwiftUI(send: @escaping (Action) -> Void) -> SwiftUI.Alert {
-    let title = Text(self.title)
-    let message = self.message.map { Text($0) }
-
-    if let primaryButton = self.primaryButton, let secondaryButton = self.secondaryButton {
-      return SwiftUI.Alert(
-        title: title,
-        message: message,
-        primaryButton: primaryButton.toSwiftUI(send: send),
-        secondaryButton: secondaryButton.toSwiftUI(send: send)
-      )
-    } else {
-      return SwiftUI.Alert(
-        title: title,
-        message: message,
-        dismissButton: self.primaryButton?.toSwiftUI(send: send)
-      )
-    }
-  }
-}
-#endif
