+++ Sources/ComposableArchitecture/UI/SwiftUI/ForEachStore.swift
