+++ Sources/ComposableArchitecture/UI/UIKit/ForEach/Store+Bindings/Store+CollectionBinding.swift
