+++ Sources/ComposableArchitecture/UI/UIKit/ForEach/Store+Bindings/Store+TableBinding.swift
